Sistemata scelta stazione diversa se piena o vuota

+++ ProgettoSistemiIntelligenti/User.cpp
@@ -59,7 +59,7 @@ void User_i::clear_added_gift(int n_s)
 	}
 }
 
-int User_i::get_value_decision()
+double User_i::get_value_decision()
 {
 	return this->value_decision;
 }
